% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_data_functions.R
\name{analyze.serosurvey}
\alias{analyze.serosurvey}
\title{analyze.serosurvey}
\usage{
analyze.serosurvey(
  data.serosurvey = data.serosurvey,
  mean = 4,
  scale.cases = 1,
  shape.controls = 1,
  scale.controls = 1,
  type = "likelihood",
  threshold = NULL,
  TNR = NULL,
  TPR = NULL,
  correct = FALSE
)
}
\arguments{
\item{data.serosurvey}{data of all observed quantitative test measures}

\item{mean}{the mean of the distribution of case sera}

\item{scale.cases}{the scale parameter of the distirbution of the case sera}

\item{shape.controls}{the shape parameter of the distribution of the control sera}

\item{scale.controls}{the scale parameter of the distribution of the control sera}

\item{type}{method used either 'naive' (cutoff-based) or 'likelihood'}

\item{threshold}{in case of 'naive' the threshold to assign someone a positive seroconversion status}

\item{correct}{TRUE if the estimate should be corrected for the sensitivity and specificity}
}
\value{
a vector with the estimate of the seroprevalence and the standard deviation in the estimate
}
\description{
analyze quantitative test measures with the likelihood-based method or one of the cutoff-based methods
}
