% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_data_functions.R
\name{AUC.ROC}
\alias{AUC.ROC}
\title{AUC.ROC}
\usage{
AUC.ROC(mean)
}
\arguments{
\item{mean}{the shape parameter of the distribution of case sera}
}
\value{
the area under the ROC-curve
}
\description{
Calculates the area under the curve of a ROC curve.
}
