% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_data_functions.R
\name{classical.threshold}
\alias{classical.threshold}
\title{classical.threshold}
\usage{
classical.threshold(mean, method = "youden")
}
\arguments{
\item{mean}{the mean of the distribution of case sera}

\item{method}{'youden' or 'specificity'}
}
\value{
a vector with the threshold value, the true positive rate and the true negative rate
}
\description{
Calculates the threshold for either the maximal youden or the high specificity method
}
