% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_data_functions.R
\name{create.figure.bootstrap}
\alias{create.figure.bootstrap}
\title{create.figure}
\usage{
create.figure.bootstrap(N_tests = 10000, N = 200, means, AUC, true_prev = 0.08)
}
\arguments{
\item{N_tests}{number of individuals tested in each simulation}

\item{N}{number of simulations performed for each AUC value}

\item{means}{the means of the titer distribution that match the AUC values}

\item{AUC}{the AUC-ROC values that should be tested}

\item{true_prev}{the true prevalence}
}
\value{
the figure
}
\description{
Make figure with inferred prevalences for all 4 methods and various AUC-ROC values.
}
