% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_data_functions.R
\name{determine.means}
\alias{determine.means}
\title{determine.means}
\usage{
determine.means(desired.AUC = c(0.7, 0.75, 0.8, 0.85, 0.9, 0.95, 0.975, 1))
}
\arguments{
\item{desired.AUC}{the AUC values that should be the result of the distribution of case sera}
}
\value{
list with as first element the true AUC-ROC values and as second element the means corresponding to those AUC values
}
\description{
Calculate the required means for the distribution of the cases to obtain the desired AUC-ROC curves
}
