% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_data_functions.R
\name{estimate.CC.distr}
\alias{estimate.CC.distr}
\title{estimate.CC.distr}
\usage{
estimate.CC.distr(
  true.mean,
  sample.sizes,
  name = "0.9",
  N_tests = 10000,
  N = 200,
  real.prev = 0.08
)
}
\arguments{
\item{sample.sizes}{the sample size that should be tested}

\item{name}{the AUC value of the tested sample sizes}

\item{N}{the number of replicates}

\item{real.prev}{the actual prevalence}

\item{mean}{mean of the distribution of case sera}
}
\value{
a dataframe with the results of all N in silico serosurveys
}
\description{
Simulates the sampling of controls and cases to estimate their distribution
}
