% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_data_functions.R
\name{get.st}
\alias{get.st}
\title{get.st}
\usage{
get.st(fit)
}
\arguments{
\item{fit}{result of the optim function}
}
\value{
matrix with the values and upper and lower bands for the
}
\description{
get.st
}
