% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_data_functions.R
\name{ll.titer}
\alias{ll.titer}
\title{ll.titer}
\usage{
ll.titer(
  prev,
  mean,
  scale.cases,
  shape.controls,
  scale.controls,
  data.serosurvey
)
}
\arguments{
\item{prev}{the seroprevalence for which the loglikelihood is calculated}

\item{mean}{the shape parameter of the distribution of the case sera}

\item{scale.cases}{the scale parameter of the distirbution of the case sera}

\item{shape.controls}{the shape parameter of the distribution of the control sera}

\item{scale.controls}{the scale parameter of the distribution of the control sera}

\item{data.serosurvey}{the reuslts of the serosurvey}
}
\value{
loglikelihood
}
\description{
Calculates the log likelihood of observing the quantitative test measures (titer.data) for a
prevalence given the distribution of quantitive test measures for the controls and cases
}
