% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_data_functions.R
\name{sim.serosurvey}
\alias{sim.serosurvey}
\title{sim.serosurvey}
\usage{
sim.serosurvey(N_tests = 10000, mean = 4, prev = 0.08)
}
\arguments{
\item{N_tests}{number of individuals enrolled in the serosurvey}

\item{mean}{shape parameter of the case sera}

\item{prev}{the seroprevalence}
}
\value{
vector with the quantitative test values of each individual
}
\description{
sim.serosurvey
}
